Divide the create form into sections
