+++ libs/query/src/lib/aa/groupCashTransfer/utils.ts
@@ -1,7 +1,6 @@
 'use client';
 import { UUID } from 'crypto';
 import { useProjectAction } from '../../projects';
-import { useSwal } from 'libs/query/src/swal';
 
 export const ACTION_NS = 'aaProject.groupCashTransfer';
 
@@ -14,15 +13,5 @@ export async function runAction(
   return q.mutateAsync({
     uuid: projectUUID as `${string}-${string}-${string}-${string}-${string}`,
     data: { action, payload },
-  });
-}
-
-export function useToast() {
-  const alert = useSwal();
-  return alert.mixin({
-    toast: true,
-    position: 'top-end',
-    showConfirmButton: false,
-    timer: 3000,
   });
 }
